add to_xrpl to Transaction
