+++ xrpl/models/transactions/transaction.py
@@ -1,14 +1,49 @@
 """The base model for all transactions and their nested object types."""
 from __future__ import annotations
 
+import re
 from dataclasses import dataclass
 from enum import Enum
 from typing import Any, Dict, List, Optional, Type, cast
 
+from xrpl.models.amounts import IssuedCurrencyAmount
 from xrpl.models.base_model import BaseModel
 from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.required import REQUIRED
 from xrpl.models.utils import require_kwargs_on_init
+
+
+def transaction_json_to_binary_codec_form(dictionary: Dict[str, Any]) -> Dict[str, Any]:
+    """
+    Returns a new dictionary in which the keys have been formatted as CamelCase and
+    standardized to be serialized by the binary codec.
+
+    Args:
+        dictionary: The dictionary to be reformatted.
+
+    Returns:
+        A new dictionary object that has been reformatted.
+    """
+    return {
+        _key_to_tx_json(key): _value_to_tx_json(value)
+        for (key, value) in dictionary.items()
+    }
+
+
+def _key_to_tx_json(key: str) -> str:
+    snaked = "".join([word.capitalize() for word in key.split("_")])
+    return re.sub(r"Id", r"ID", snaked)
+
+
+def _value_to_tx_json(value: Any) -> Any:
+    # IssuedCurrencyAmount is a special case and should not be snake cased
+    if IssuedCurrencyAmount.is_dict_of_model(value):
+        return {key: _value_to_tx_json(sub_value) for (key, sub_value) in value.items()}
+    if isinstance(value, dict):
+        return transaction_json_to_binary_codec_form(value)
+    if isinstance(value, list):
+        return [_value_to_tx_json(sub_value) for sub_value in value]
+    return value
 
 
 class TransactionType(str, Enum):
@@ -210,6 +245,16 @@ class Transaction(BaseModel):
         # we need to override this because transaction_type is using ``field``
         # which will not include the value in the objects __dict__
         return {**super().to_dict(), "transaction_type": self.transaction_type.value}
+
+    def to_xrpl(self: Transaction) -> Dict[str, Any]:
+        """
+        Creates a JSON-like dictionary in the JSON format used by the binary codec
+        based on the Transaction object.
+
+        Returns:
+            A JSON-like dictionary in the JSON format used by the binary codec.
+        """
+        return transaction_json_to_binary_codec_form(self.to_dict())
 
     @classmethod
     def from_dict(cls: Type[Transaction], value: Dict[str, Any]) -> Transaction:
